removed url and deleted paginatedTable.jsx
