--- packages/react-core/src/demos/sampleData2.tsx
@@ -1,159 +1,0 @@
-export interface SampleRowData {
-  status: string;
-  location: string;
-  lastModified: string;
-  url: string;
-}
-
-export const rows: SampleRowData[] = [
-  {
-    status: 'running',
-    location: 'raleigh',
-    lastModified: 'oneHr',
-    url: 'http://www.redhat.com/en/office-locations/node1'
-  },
-  {
-    status: 'stopped',
-    location: 'boston',
-    lastModified: 'threeHrs',
-    url: 'http://www.redhat.com/en/office-locations/node2'
-  },
-  {
-    status: 'down',
-    location: 'atlanta',
-    lastModified: 'fiveHrs',
-    url: 'http://www.redhat.com/en/office-locations/node3'
-  },
-  {
-    status: 'needsMaintenance',
-    location: 'sanFrancisco',
-    lastModified: 'sevenMins',
-    url: 'http://www.redhat.com/en/office-locations/node4'
-  },
-  {
-    status: 'running',
-    location: 'raleigh',
-    lastModified: 'fortyTwoMins',
-    url: 'http://www.redhat.com/en/office-locations/node5'
-  },
-  {
-    status: 'stopped',
-    location: 'boston',
-    lastModified: 'twoDays',
-    url: 'http://www.redhat.com/en/office-locations/node6'
-  },
-  {
-    status: 'down',
-    location: 'atlanta',
-    lastModified: 'oneMonth',
-    url: 'http://www.redhat.com/en/office-locations/node7'
-  },
-  {
-    status: 'needsMaintenance',
-    location: 'sanFrancisco',
-    lastModified: 'oneHr',
-    url: 'http://www.redhat.com/en/office-locations/node8'
-  },
-  {
-    status: 'running',
-    location: 'raleigh',
-    lastModified: 'threeHrs',
-    url: 'http://www.redhat.com/en/office-locations/node9'
-  },
-  {
-    status: 'stopped',
-    location: 'boston',
-    lastModified: 'fiveHrs',
-    url: 'http://www.redhat.com/en/office-locations/node10'
-  },
-  {
-    status: 'down',
-    location: 'atlanta',
-    lastModified: 'sevenMins',
-    url: 'http://www.redhat.com/en/office-locations/node11'
-  },
-  {
-    status: 'needsMaintenance',
-    location: 'sanFrancisco',
-    lastModified: 'fortyTwoMins',
-    url: 'http://www.redhat.com/en/office-locations/node12'
-  },
-  {
-    status: 'running',
-    location: 'raleigh',
-    lastModified: 'twoDays',
-    url: 'http://www.redhat.com/en/office-locations/node13'
-  },
-  {
-    status: 'stopped',
-    location: 'boston',
-    lastModified: 'oneMonth',
-    url: 'http://www.redhat.com/en/office-locations/node14'
-  },
-  {
-    status: 'down',
-    location: 'atlanta',
-    lastModified: 'oneHr',
-    url: 'http://www.redhat.com/en/office-locations/node15'
-  },
-  {
-    status: 'needsMaintenance',
-    location: 'sanFrancisco',
-    lastModified: 'threeHrs',
-    url: 'http://www.redhat.com/en/office-locations/node16'
-  },
-  {
-    status: 'running',
-    location: 'raleigh',
-    lastModified: 'fiveHrs',
-    url: 'http://www.redhat.com/en/office-locations/node17'
-  },
-  {
-    status: 'stopped',
-    location: 'boston',
-    lastModified: 'sevenMins',
-    url: 'http://www.redhat.com/en/office-locations/node18'
-  },
-  {
-    status: 'down',
-    location: 'atlanta',
-    lastModified: 'fortyTwoMins',
-    url: 'http://www.redhat.com/en/office-locations/node19'
-  },
-  {
-    status: 'needsMaintenance',
-    location: 'sanFrancisco',
-    lastModified: 'twoDays',
-    url: 'http://www.redhat.com/en/office-locations/node20'
-  },
-  {
-    status: 'running',
-    location: 'raleigh',
-    lastModified: 'oneMonth',
-    url: 'http://www.redhat.com/en/office-locations/node21'
-  },
-  {
-    status: 'stopped',
-    location: 'boston',
-    lastModified: 'oneHr',
-    url: 'http://www.redhat.com/en/office-locations/node22'
-  },
-  {
-    status: 'down',
-    location: 'atlanta',
-    lastModified: 'threeHrs',
-    url: 'http://www.redhat.com/en/office-locations/node23'
-  },
-  {
-    status: 'needsMaintenance',
-    location: 'sanFrancisco',
-    lastModified: 'fiveHrs',
-    url: 'http://www.redhat.com/en/office-locations/node24'
-  },
-  {
-    status: 'running',
-    location: 'raleigh',
-    lastModified: 'sevenMins',
-    url: 'http://www.redhat.com/en/office-locations/node25'
-  }
-];
