+++ packages/react-core/src/demos/sampleDataRTL.tsx
@@ -1,0 +1,133 @@
+export interface SampleRowData {
+  status: string;
+  location: string;
+  lastModified: string;
+}
+
+export const rows: SampleRowData[] = [
+  {
+    status: 'running',
+    location: 'raleigh',
+    lastModified: 'oneHr'
+  },
+  {
+    status: 'stopped',
+    location: 'boston',
+    lastModified: 'threeHrs'
+  },
+  {
+    status: 'down',
+    location: 'atlanta',
+    lastModified: 'fiveHrs'
+  },
+  {
+    status: 'needsMaintenance',
+    location: 'sanFrancisco',
+    lastModified: 'sevenMins'
+  },
+  {
+    status: 'running',
+    location: 'raleigh',
+    lastModified: 'fortyTwoMins'
+  },
+  {
+    status: 'stopped',
+    location: 'boston',
+    lastModified: 'twoDays'
+  },
+  {
+    status: 'down',
+    location: 'atlanta',
+    lastModified: 'oneMonth'
+  },
+  {
+    status: 'needsMaintenance',
+    location: 'sanFrancisco',
+    lastModified: 'oneHr'
+  },
+  {
+    status: 'running',
+    location: 'raleigh',
+    lastModified: 'threeHrs'
+  },
+  {
+    status: 'stopped',
+    location: 'boston',
+    lastModified: 'fiveHrs'
+  },
+  {
+    status: 'down',
+    location: 'atlanta',
+    lastModified: 'sevenMins'
+  },
+  {
+    status: 'needsMaintenance',
+    location: 'sanFrancisco',
+    lastModified: 'fortyTwoMins'
+  },
+  {
+    status: 'running',
+    location: 'raleigh',
+    lastModified: 'twoDays'
+  },
+  {
+    status: 'stopped',
+    location: 'boston',
+    lastModified: 'oneMonth'
+  },
+  {
+    status: 'down',
+    location: 'atlanta',
+    lastModified: 'oneHr'
+  },
+  {
+    status: 'needsMaintenance',
+    location: 'sanFrancisco',
+    lastModified: 'threeHrs'
+  },
+  {
+    status: 'running',
+    location: 'raleigh',
+    lastModified: 'fiveHrs'
+  },
+  {
+    status: 'stopped',
+    location: 'boston',
+    lastModified: 'sevenMins'
+  },
+  {
+    status: 'down',
+    location: 'atlanta',
+    lastModified: 'fortyTwoMins'
+  },
+  {
+    status: 'needsMaintenance',
+    location: 'sanFrancisco',
+    lastModified: 'twoDays'
+  },
+  {
+    status: 'running',
+    location: 'raleigh',
+    lastModified: 'oneMonth'
+  },
+  {
+    status: 'stopped',
+    location: 'boston',
+    lastModified: 'oneHr'
+  },
+  {
+    status: 'down',
+    location: 'atlanta',
+    lastModified: 'threeHrs'
+  },
+  {
+    status: 'needsMaintenance',
+    location: 'sanFrancisco',
+    lastModified: 'fiveHrs'
+  },
+  {
+    status: 'running',
+    location: 'raleigh',
+    lastModified: 'sevenMins'
+  }
+];
